requirements.txt
ftfy~=5.8
regex~=2021.8.3
tqdm~=4.63.0
torch
torchvision~=0.11.3

pillow~=8.3.1
numpy~=1.19.2
pip~=21.2.2
dataclasses~=0.8
attrs~=21.4.0
zlib~=1.2.13
wheel~=0.37.1
openssl~=1.1.1u
tornado~=6.1
jinja2~=3.0.3
setuptools~=58.0.4
packaging~=21.3
pyparsing~=3.0.4
zipp~=3.6.0
typing_extensions~=4.1.1
pytest~=7.0.1
clip~=1.0
configparser~=5.2.0
pathlib~=1.0.1
future~=0.18.3
matplotlib~=3.3.4
scikit-image~=0.17.2
ipython~=7.16.1
requests~=2.27.1
wget~=3.2
yaml~=0.2.5
pyyaml~=5.4.1
scikit-learn~=0.24.2